Переработка аннотаций

+++ datacls.py
@@ -21,8 +21,8 @@ class DiplomaParametrs:
 
 @dataclass()
 class Point:
-    x: int = 0
-    y: int = 0
+    x: float = 0
+    y: float = 0
 
 @dataclass()
 class DiplomaTitleLayoutParametrs:
